Add auxiliary files
Adds configuration files and systemd units. Files for the backup server
are under backup/. Files for the node are under node/. Both directories
emulate the filesystem structure of the respective target.

Files are symlinked rather than duplicated or hard-linked. If copying
the directories to the root filesystem of the target, dereference the
symlinks.