Run update scripts using Bash instead of sh(1)
As this system is designed to be maintainable by someone with limited
shell scripting or system administration skills, it's a better idea to
use Bash for update scripts, as many tutorials for shell scripting are
actually using Bash syntax, not POSIX sh syntax.

This should reduce errors caused by updates.