Generate provision.sh by including other files
Uses the GNU M4 macro processor to process the file `provision.sh.in'
and turn it into `provision.sh' by expanding macros. Currently, the only
macros used are to include the various systemd unit files.

This helps avoid having to make a change to the systemd unit file as
well as the contents of that file within the provision script. Simply
updating the unit file is enough, as its contents will be included in
the resulting script.

A Makefile is provided to make processing the provisioning script easy.
Just run `make' to process the script into its final form.