Fix backup on shutdown functionality
Starting a unit before shutdown.target doesn't work. Instead, this
change starts a service that does when the system boots. This service
runs the client script when it is stopped, effectively backing up the
system on shutdown.

This also requires a change to the configuration file format for time
entries.