client.py: Use a lock file to prevent double-backups
The backup-on-shutdown systemd service will run a backup when it is
stopped. This means that if the configuration file changes and backups
on shutdown are disabled, that service will be stopped by the client,
thus triggering a backup while the current one is already happening.

To prevent this, we make the client script obtain a lock on a file. If
the script cannot obtain the lock, it dies. This ensures that only one
instance of the client script will run at a time.

The lock file defaults to /var/run/backup_client.lock. If the
RUNTIME_DIRECTORY environment variable is set, the lock file is
$RUNTIME_DIRECTORY/backup_client.lock.