client.py: Allow skipping updates from command line
Adds the command-line parameter --resume-after-update <EPOCH>. This
parameter will make the client script skip application of updates with
an epoch less than or equal to <EPOCH>.

This is not intended to be used manually. It will be used by the client
script in order to re-execute itself after each update is applied.