client.py: Re-execute client.py after updates
Each update may want to modify the client script. In order to support
this, we must re-execute the client script after each update is applied.

This commit implements that. After each update is applied, the client
script is re-executed. The "--resume-after-update <EPOCH>" argument is
given, so the new instance of the backup client won't apply the same
update again.