provision.sh: Remove old artifacts when installing
This commit changes the provisioning script to remove directories left
over from previous installations that could cause unwanted behavior.
This includes:
  - Systemd unit override directories
  - State directories (e.g. stored configuration)
  - Static data directories

The main reason for this was that installing a new client script with a
different backup server IP address didn't work, since the IP address
from the old configuration was used anyways. This was fixed in commit
[COMMIT], so it shouldn't be a problem anymore.