Fix binary paths
On systems with separate /bin and /usr/bin directories, true(1) is
located in /bin/true, and calling it as /usr/bin/true will fail. This is
in accordance with the Filesystem Hierarchy Standard [1].

On Debian 10 and earlier, systemctl(1) is installed in /bin rather than
/usr/bin where it is supposed to be. To avoid problems, this commit
makes the client script call systemctl(1) without an absolute path and
relies on execvp(2) to figure it out.

[1]: https://refspecs.linuxfoundation.org/FHS_3.0/fhs/ch03s04.html#requirements2