find_missing.py: Implement email notifications
This commit adds the `--email' command-line flag to the find_missing.py
script. When present, the script will send an email with its results to
the emails listed in the `EMAIL_RECIPIENTS' list in the script.

Email and SMTP parameters may be specified using environment variables
to avoid leaking sensitive information in the script.